+++ allennlp_models/classification/tango/__init__.py
@@ -1,0 +1,1 @@
+from allennlp_models.classification.tango.imdb import ImdbInstances